+++ src/App.js
@@ -2,10 +2,10 @@ import React, { Component } from 'react';
 import logo from './nearForm-logo.svg';
 import './App.css';
 import { Connector } from 'mqtt-react';
-import _MessageContainer from './MessageContainer.js';
+import _MessageContainer from './components/Class/MessageContainer.js';
 import {subscribe} from 'mqtt-react';
 import API from './API.js';
-import HookMqtt from './indx.js';
+import HookMqtt from './components/Hook/';
 import Amplify from 'aws-amplify';
 import awsconfig from './aws-exports';
 import {AmplifySignOut, withAuthenticator} from '@aws-amplify/ui-react';
