+++ src/components/Hook/Subscriber.js
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
 import { Card, Form, Input, Row, Col, Button, Select } from 'antd';
-import { QosOption } from './indx'
+import { QosOption } from './index'
 
 const Subscriber = ({ sub, unSub, showUnsub }) => {
   const [form] = Form.useForm();
